+++ pattern_dict.py
@@ -1,0 +1,4 @@
+def pattern_dict():
+   print("Coming soon... Returning to menu :)")
+   print()
+   menu()
